Update template, add more logging, fix 404s dying
Whereas previously the handler was throwing errors via log.Fatal, that's
not what we want to happen. So now they just print them and return,
waiting for someone else's cool middleware stack to take the burden off
of our error-filled hands.

Also add a title to the page that is equivalent to the current dir. In
the case of ".", it changes to "Current".